+++ test/java/io/File/DeleteOnExit.java
@@ -31,10 +31,10 @@
 import java.io.File;
 
 public class DeleteOnExit  {
-    
-    static String tmpdir = System.getProperty("java.io.tmpdir");    
-    static String java = System.getProperty("java.home") + File.separator +    
-	                 "bin" + File.separator + "java";
+
+    static String tmpdir = System.getProperty("java.io.tmpdir");
+    static String java = System.getProperty("java.home") + File.separator +
+                         "bin" + File.separator + "java";
     static File file1 = new File(tmpdir + "deletedOnExit1");
     static File file2 = new File(tmpdir + "deletedOnExit2");
     static File file3 = new File(tmpdir + "deletedOnExit3");
@@ -45,60 +45,60 @@ public class DeleteOnExit  {
     static File file5 = new File(dir + File.separator + "dxnsdnguidfgejngognrogn");
     static File file6 = new File(dir + File.separator + "mmmmmmsdmfgmdsmfgmdsfgm");
     static File file7 = new File(dir + File.separator + "12345566777");
-    
+
     public static void main (String args[]) throws Exception{
         if (args.length == 0) {
             Runtime.getRuntime().exec(java +  " DeleteOnExit -test").waitFor();
             if (file1.exists() || file2.exists() || file3.exists() ||
-		dir.exists()   || file4.exists() || file5.exists() ||
-		file6.exists() || file7.exists())  {
+                dir.exists()   || file4.exists() || file5.exists() ||
+                file6.exists() || file7.exists())  {
 
-		System.out.println(file1 + ", exists = " + file1.exists());
-		System.out.println(file2 + ", exists = " + file2.exists());
-		System.out.println(file3 + ", exists = " + file3.exists());
-		System.out.println(dir + ", exists = " + dir.exists());
-		System.out.println(file4 + ", exists = " + file4.exists());
-		System.out.println(file5 + ", exists = " + file5.exists());
-		System.out.println(file6 + ", exists = " + file6.exists());
-		System.out.println(file7 + ", exists = " + file7.exists());
+                System.out.println(file1 + ", exists = " + file1.exists());
+                System.out.println(file2 + ", exists = " + file2.exists());
+                System.out.println(file3 + ", exists = " + file3.exists());
+                System.out.println(dir + ", exists = " + dir.exists());
+                System.out.println(file4 + ", exists = " + file4.exists());
+                System.out.println(file5 + ", exists = " + file5.exists());
+                System.out.println(file6 + ", exists = " + file6.exists());
+                System.out.println(file7 + ", exists = " + file7.exists());
 
-		// cleanup undeleted dir if test fails
-	        dir.delete();
+                // cleanup undeleted dir if test fails
+                dir.delete();
 
                 throw new Exception("File exists");
-	    }
+            }
         } else {
             file1.createNewFile();
             file2.createNewFile();
-	    file3.createNewFile();
+            file3.createNewFile();
             file1.deleteOnExit();
             file2.deleteOnExit();
             file3.deleteOnExit();
 
-	    // verify that deleting a File marked deleteOnExit will not cause a problem 
-	    // during shutdown.
-	    file3.delete();
+            // verify that deleting a File marked deleteOnExit will not cause a problem
+            // during shutdown.
+            file3.delete();
 
-	    // verify that calling deleteOnExit multiple times on a File does not cause
-	    // a problem during shutdown.
-	    file2.deleteOnExit();
-	    file2.deleteOnExit();
-	    file2.deleteOnExit();
+            // verify that calling deleteOnExit multiple times on a File does not cause
+            // a problem during shutdown.
+            file2.deleteOnExit();
+            file2.deleteOnExit();
+            file2.deleteOnExit();
 
-	    // Verify DeleteOnExit Internal implementation deletion order.
-	    if (dir.mkdir()) {
-		dir.deleteOnExit();
+            // Verify DeleteOnExit Internal implementation deletion order.
+            if (dir.mkdir()) {
+                dir.deleteOnExit();
 
-		file4.createNewFile();
-		file5.createNewFile();
-		file6.createNewFile();
-		file7.createNewFile();
+                file4.createNewFile();
+                file5.createNewFile();
+                file6.createNewFile();
+                file7.createNewFile();
 
-		file4.deleteOnExit();
-		file5.deleteOnExit();
-		file6.deleteOnExit();
-		file7.deleteOnExit();
-	    }
-        }        
+                file4.deleteOnExit();
+                file5.deleteOnExit();
+                file6.deleteOnExit();
+                file7.deleteOnExit();
+            }
+        }
     }
 }
